Create Main-Program
This paper derives a general formula applicable to the sum of the first n positive integers raised to any natural number power p greater than three. Initially focusing on a novel geometric derivation for the sum of the first n positive integers raised to the power of 3, dimensional induction, in conjunction with theoretical analyses, is conducted to extend the proof to the summation of integers raised to powers of 4 or greater. Through leveraging this geometric method in higher dimensions, we establish a recursive expression that holds true for all natural numbers p greater than 3. By calculating the average runtimes of the researched expression and Faulhaber’s formula, we found that the researched formula was able to greatly improve the efficiency of calculation for summations raised to higher powers of p, as, unlike the Faulhaber formula, whose runtimes increased proportionally to the value of p, it was able to maintain consistently similar runtimes as p increased.